Rename put scripts into scripts
